oc-mail: Fix crash on syncing deleted items
It crashes because we are returning deleted fmids without a
valid change number in sogo_folder_get_deleted_fmids.

This was due in the synchroniseCache process during the
process of vanished items, the deleted message was not
available in versions plist and there were no
SyncLastDeleteChangeNumber so we need to create the entry
with a change number to avoid this problem.

We can reproduce it by creating a profile in Outlook
, close it, create and delete messages using an IMAP client,
and finally reopen the Outlook.