oc-calendar: Store HTML from event description
According to [MS-OXCICAL] Section 2.1.3.1.1.20.27
it is stored in X-ALT-DESC tag with FMTTYPE attribute
set to "text/html". With that, we can export it as
ics and import properly in other Outlook without a problem.

This makes current supported RTF commands be maintained
between Outlook clients provided that in SOGo Webmail
plain text is only shown.