oc-calendar: Do not skip submit of meeting request and response messages
When the PidTagMessageClass is IPM.Schedule.Meeting.*, it is ignored
as only messages with class IPM.Note or IPM.Sharing were taken into account.

There was a comment saying that the purpose of the check is to skip
invitation replies, but with no apparent reason for that, or at least
not explained why.

Not sure why in some cases this has worked without this change, but at least
now seems pretty easy to reproduce it, and this seems to fix the issue.