oc-calendar: Get PidLidAppointmentSequence from invitation mails
This property didn't have a getter in MAPIStoreMailMessage, and the
client copies it from the update mail into the calendar appointment in
order to keep track of the appointment version. Failing to return it
caused the event not to be saved because the sequence was lower than
the previous version of the event.