oc-calendar: Return PidLidMeetingType more accurately
This property ([MS-OXOCAL] 2.2.6.5, 2.2.8.6) specifies the type of the
update, invitation or cancellation objects, and the client uses it to
adjust how the event notification mails are displayed.

We were returning 0x1 (new event invitation) before. Now, it can also
indicate that we're dealing with an update that requires the attendee's
validation. This is returned at the MAPIStoreMailMessage level because
we need the header to discern in which situation we are and because it
is a property that belongs to notification mails, not events.