oc-calendar: Deal with soft-deleted events when updating them
When updating an event, Oultook deletes the appointment and stores a
copy in the calendar. The SOGo/OC layer tries to recover the old
instance of the event when we're saving the copy, associating the URL of
the initial event to the ID of our new object. When that URL belonged to
an object that had been soft-deleted, it wasn't possible to register the
new ID, so the appointment disappeared from the calendar in Outlook.