oc-calendar: Don't send declination mail when updating an event
When we accept an event update from Outlook, it creates a copy of the
event based on the notification mail, deletes the appointment in the
calendar and stores the copy. When we prepare the deletion of an
appointment, SOGo sends a cancellation/declination mail.

This patch skips this last part when we detect that the update request
hasn't been responded.