oc-calendar: Cancel events when sending the cancellation mail
SOGo/OC doesn't send the event notification mails (which include e.g.
invitations, cancellations or update confirmations) because this is
handled at the SOGo/Appointments layer. This patch detects when a mail
is a declination/cancellation one and prepares the deletion of the
event, which mainly means that it sends the cancellation mail.

In addition, it detects if the original invitation is being decline and,
in this case, it deletes the appointment object (Outlook hasn't stored
it yet, so it won't try to delete it).